+++ src/features/transfer-ownership/FormSteps/index.ts
@@ -2,4 +2,3 @@ export * from './Complete';
 export * from './Confirm';
 export * from './Details';
 export * from './Transaction';
-export * from './types';
