--- src/features/transfer-ownership/TransferOwnership.module.scss
@@ -1,16 +1,0 @@
-@import '../../../node_modules/@zero-tech/zui/styles/_colors.scss';
-@import '../../../node_modules/@zero-tech/zui/styles/_typography.scss';
-
-.Container {
-	width: 36.25rem;
-	padding: 3rem;
-
-	.ErrorMessage {
-		display: block;
-		color: $helitrope;
-		font-size: $size-small;
-		font-weight: $font-weight-bold;
-		margin-top: 0.75rem;
-		text-align: center;
-	}
-}
