+++ src/features/transfer-ownership/TransferOwnershipForm/TransferOwnershipForm.constants.ts
@@ -1,0 +1,7 @@
+export enum Step {
+	DETAILS,
+	CONFIRM,
+	TRANSACTION_APPROVAL,
+	TRANSACTION_IN_PROGRESS,
+	COMPLETE,
+}
