+++ src/features/transfer-ownership/TransferOwnershipForm/index.ts
@@ -1,0 +1,1 @@
+export * from './TransferOwnershipForm';
