+++ src/features/transfer-ownership/TransferOwnershipModal/TransferOwnershipModal.module.scss
@@ -1,0 +1,4 @@
+.Container {
+	width: 36.25rem;
+	padding: 3rem;
+}
