+++ src/features/transfer-ownership/TransferOwnershipModal/index.ts
@@ -1,0 +1,1 @@
+export * from './TransferOwnershipModal';
